+++ tests/integration/run_tests.py
@@ -34,6 +34,7 @@ Built for 1) manual use and 2) continuous integration.
 """
 import argparse
 import shutil
+import sys
 import time
 from contextlib import suppress
 from subprocess import PIPE, Popen
@@ -76,6 +77,12 @@ def run_single_test(name, test, test_label="", buffer_length=50):
         if passed is False:
             info["violated_condition"] = (condition, i, len(conditions))
             break
+
+    if len(test) == 4:
+        end_process = Popen(test[3], stdout=PIPE, stderr=PIPE, shell=True)
+        end_stdout, end_stderr = end_process.communicate()
+        info["end_stdout"] = end_stdout
+        info["end_stderr"] = end_stderr
 
     return passed, info
 
@@ -138,7 +145,11 @@ def run_tests(args, tests):
         n_to_run = 0
         selective = True
         for test_id, test in enumerate(tests.values()):
-            if len(test) == 3 and test[2] == "local":
+            # Ensures that test definitions are atleast size 3.
+            # 'local' variable is stored in 3rd element of the test definitions,
+            # but an optional 4th element can be provided for an ending command
+            # to be ran after all checks have been made.
+            if len(test) >= 3 and test[2] == "local":
                 args.ids.append(test_id + 1)
                 n_to_run += 1
 
@@ -210,7 +221,7 @@ def main():
 
     clear_test_studies_dir()
     result = run_tests(args, tests)
-    return result
+    sys.exit(result)
 
 
 if __name__ == "__main__":